seeds/requirements.txt
faker
psycopg2-binary==2.9.10